fixed bugs in db and restoring of match match restore now works by setting a env variable in the docker container

+++ app/endpoints/csgo_events.py
@@ -1,5 +1,7 @@
 import json
 import logging
+import os
+import random
 from typing import Dict
 from time import sleep
 
@@ -8,8 +10,11 @@ from fastapi import Request, Depends
 
 from endpoints import csgo_stats_event, auth_api
 from rcon import RCON
+from servers import ServerManager
 from sql import db
 from elo import calculate_elo
+
+server_manger: ServerManager = None
 
 
 def going_live(event: Dict):
@@ -35,13 +40,19 @@ def demo_upload_ended(event: Dict):
     match.finished = 2
     match.update_attribute("finished")
 
-    res = requests.delete(f"http://{server.ip}/api/match", json={"id": server.id}, timeout=60,
-                          headers=auth_api.login_to_master_headers())
-    if res.status_code == 200:
-        logging.info(
-            f"After demo has been uploaded, container running matchid: {match.matchid}, (name={server.container_name}) has been stopped")
+    if os.getenv("EXTERNAL_IP", "127.0.0.1") != server.ip and server.ip != "host.docker.internal":
+        logging.info(f"Shutting down remote container: server={server.ip}, match_id={match.matchid}")
+        res = requests.delete(f"http://{server.ip}/api/match", json={"id": server.id}, timeout=60,
+                              headers=auth_api.login_to_master_headers())
+
+        if res.status_code == 200:
+            logging.info(
+                f"After demo has been uploaded, container running matchid: {match.matchid}, (name={server.container_name}) has been stopped")
+        else:
+            logging.error(f"Unable to stop container for match: {match}, server: {server} -> {res.text}")
     else:
-        logging.error(f"Unable to stop container for match: {match}, server: {server} -> {res.text}")
+        logging.info(f"Shutting down local container: server={server.ip}, match_id={match.matchid}")
+        server_manger.stop_match(server.id)
 
 
 def map_result(event: Dict):
@@ -57,6 +68,9 @@ def map_result(event: Dict):
 
     team1 = db.get_team_by_id(match.team1)
     team2 = db.get_team_by_id(match.team2)
+
+    team1_elo = team1.elo
+    team2_elo = team2.elo
     logging.info(f"Updating ELO: {team1.elo}, {team2.elo}, {event['team1_score']}, {event['team2_score']}")
 
     team1.elo, team2.elo = calculate_elo(team1.elo, team2.elo, event["team1_score"], event["team2_score"])
@@ -64,6 +78,12 @@ def map_result(event: Dict):
     team2.update_attribute("elo")
 
     logging.info(f"Updated ELO: {team1.elo}, {team2.elo}, {event['team1_score']}, {event['team2_score']}")
+
+    server: db.Server = db.get_server_for_match(match.matchid)
+    with RCON(server.ip, server.port, "pass") as rconn:
+        rconn.exec_command("say ELO updated:")
+        rconn.exec_command(f"say {team1.name}: {team1.elo}, diff: {team1.elo - team1_elo}")
+        rconn.exec_command(f"say {team2.name}: {team2.elo}, diff: {team2.elo - team2_elo}")
 
 
 def series_end(event: Dict):
@@ -85,7 +105,6 @@ def player_say(event: Dict):
         with RCON(server.ip, server.port, "pass") as rconn:
             rconn.exec_command("say spinning")
             rconn.exec_command("say done")
-            # TODO add random messages
 
 
 callbacks = {
@@ -120,3 +139,8 @@ def set_api_routes(app):
             callbacks[event["event"]](event)
 
         return ""
+
+
+def set_server_manager(server_manager: ServerManager):
+    global server_manger
+    server_manger = server_manager
